pyproject.toml
[project]
name = "event-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.0",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-groq>=1.0.1",
    "langchain-huggingface>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]
